+++ homebrew/maccrab.rb
@@ -1,5 +1,5 @@
 cask "maccrab" do
-  version "1.6.19"
+  version "1.6.20"
   sha256 "9077a99014b32a5f795122688b9a5a23502932e9111ec5ec57ce23dc720f2121"
 
   url "https://github.com/peterhanily/maccrab/releases/download/v#{version}/MacCrab-v#{version}.dmg"